Complete Overhaul From Original Project
Long story short, the transition from Eclipse to Xcode made it an
absolute pain for me to figure out, and on top of that had a *ton* of
errors due to keyboard not properly reading my keystrokes as well as
missing large chunks of the assignment. Because of this, I’ve made the
decision to re-do everything and delete the previous project.